Applied patch from SourceForge (#1509727, Date Submitted: 2006-06-21 07:49, Date Last Updated: 2006-07-06 18:51):
The OpenCBM module in cbm4linux uses the MOD_PARM()
macro, which is deprecated in older 2.6.x kernels, and
obsoleted (removed) in newer 2.6.x kernels.

This results in the following compilation error:

error: expected ')' before string constant

in cbm_module.c, lines 111, 115, 118, 126.

The attached patch fixes this problem, while remaining
compatible with older kernels (2.4.x and older). This
patch is part of v0.4.0.17.

- Spiro.