Applied patch from SourceForge (#1742142, Date Submitted: 2007-06-23 20:10, Date Last Updated: 2007-06-23 20:10):
Normal users are not allowed to access OpenCBM / udev issue

If you are running a kernel with udev support, any normal user does not get
access to OpenCBM, as the rights of /dev/cbm are wrong. The attached patch
makes sure anyone from group "users" gets access to the OpenCBM driver.
Replace "users" with any other group in
sys/linux/45-opencbm-plugin-xa1541.rules if you want to give this
permission to another group.

Spiro