Applied patch from SourceForge (#1772084, Author: Frederic Briere, Date Submitted: 2007-08-11 05:36, Date Last Updated: 2007-08-11 05:36):
High-bit character in d64copy and libcbmcopy

There is a high-bit character (0xB4) in the help/manpage of d64copy, and in
a debug message of libcbmcopy. This is meant to be an apostrophe, but
doesn't look as such to those of us who left the blissful world of
ISO-8859-1.

I'm attaching a patch that converts them to ASCII apostrophes. There's a
lot more of them floating around (in the HTML doc and libcommon, mostly),
but I didn't dare messing with that.