Applied patch from SourceForge (#1805377, Author: Frederic Briere, Date Submitted: 2007-10-01 03:28, Date Last Updated: 2007-10-01 03:28):
struct pt_regs * argument was removed in Linux 2.6.19

The third argument to the interrupt callback function, of type struct
pt_regs *, was removed in Linux 2.6.19 (my understanding is that nobody was
using it).

This patch removes the corresponding useless argument in cbm_interrupt(),
and adjusts the declaration of cbm_interrupt_pp() according to the kernel
version.