Applied patch from SourceForge (#1772070, Author: Frederic Briere, Date Submitted: 2007-08-11 04:17, Date Last Updated: 2007-08-11 06:18):
libopencbm.a does not include libarch

The lib Makefile only brings libarch.a in for the shared library, and not
for the static library.

The attached patch builds two libarch'es (one with PIC, one without), and
pulls in each one when needed. The second half is, for lack of a better
word, ugly, but I couldn't find any other way to append an .a to another.

This patch is meant to be applied *instead* of your
previous shlib patch, and not on top of it.