Replaced the old manpages with new ones copied from the manual
Here's a new set of manpages, copied from the manual and sprinkled with
nroff markup.

I took the liberty of making a couple of modifications along the way;
mostly having to do with the synopsis, the order of options, replacing
"disc" with "disk", and maybe fixing a typo or two. Nevertheless, I tried
not to stray too far from the original, so as to make it easier to keep
them synchronized in the future.