Linux module: Style: Re-indent cbm_module.[ch] according to Linux standards
This patch is part of a series that brings the coding style closer to
that of the Linux kernel, as documented in CodingStyle.

This patch was produced by running Linux's scripts/Lindent on
cbm_module.[ch], and fixing the two erroneous "loff_t * ppos" that crept
in, as well as re-indenting the rcsid line.