Add new board: ZoomFloppy. This will be a commercially-available design based on the ATMEGA32U2. It is not compiled yet until I update my toolchain to support this new CPU. Add a schematic for this board.
Add new design based on ZoomFloppy using the Bumble-B v1, which uses the
AT90USB162. It is identical at the board level to the ZoomFloppy.

Update the protocol to v6. The xum1541 plugin will also need to be rebuilt.
This protocol makes all commands handle aborts cleanly while still maintaining
infinite listener holdoff. This means you can start a command like "cbmctrl
change 8" and leave it for hours, abort it, and restart it cleanly. Also,
by making the cbm_raw_read/write() protocols inline, we save some small
time on buffering data 128-bytes at a time in the xum1541's RAM. This was
leftover from the xu1541 and was originally needed to deal with getting off
the bus fast while letting the drive take a while to answer. With our
blocking bulk commands, we don't need that any more.

The makefile now builds all firmware versions in one pass.

This code was tested with both the USBKEY and ZoomFloppy board layouts
on Windows.