Fix parallel port test failure and hang on some empty tracks. Both were rooted in the optimization of IO routines revealing some wrong assumptions about timing.
The par port test failed by reading doubles of the pattern (00, 00, 01, 01...)
This was because the optimized IO started a second read too quickly.
This only showed up in the par port test because otherwise the drive is
even slower in returning bytes. The fix is to delay 5 us between each
handshaked read.

The hang on empty tracks was caused by the handshaked read timing out.
Because of the optimized IO, the counter hit zero too quickly when the
drive code took longer to respond. The fix is to wait forever and trust
in the watchdog timer to catch us if we truly get hung up.