Fix timing of IEC write to match the timing diagrams as well as testing. Since I optimized the IO routines, the timing of this is more critical. Add delays in appropriate places, named according to the spec. For example, ATN response time (Tat) -> IEC_T_AT.
For the 1 ms delays (Tat and Tf), we actually allow up to 2 ms.

Eac time we send a byte, we delay the Tbb of 100 us. In the case of the
last byte sent under ATN or a talk-atn turn around, this is a bit extra
since Tr (20 us) is all that's required after the device acknowledges the
byte (listener data accepted).

Each time we change an IO line, delay 0.5 us before polling it. I have
measured on a scope that this is how long it takes a line to transition
based on the pullups (0 to 1).