Cache up to 4 bytes of nib writes instead of just the first one. nibwrite sends 2 bytes (ihs and sync) to the 1571 when starting to write a track whereas it only sends 1 byte for the 1541. This fixes nibwrite, which didn't work on the 1571.
Background:

Because of the USB latency, we have to consume the handshake byte that is
sent from the PC immediately before the transfer of 8192 track bytes.

nib_parburst_write_checked() keeps a state machine that looks for
parallel commands (0x00, 0x55, 0xaa, 0xff, ...) If we see a read or
write track command, we set a flag suppressNibCmd that means "don't
immediately send the handshake, the read/write track routine will do it."

The fix will now cache the 1 byte (1541) or 2 bytes (1571) sent from the
PC and send it to the 1571 in conjunction with the write track command
when it arrives. That way all data has been received over USB before
initiating the transfer with the drive.