Further protocol fixes to cbmcopy, regarding protocols s1 and pp:
In s1 protocol, when writing and being at the end of a transfer,
the host side (PC) must not wait for the device to send a final
signaling on DATA, because the 6502 code does not explicitly send
this. When the previous CLOCK signaling was send from the host side
to the device, the device routine returns from the "noerror" part
of the transfer routine.

In pp protocol, for both directions, writing as well as reading,
the host side (PC) must not wait for a final handshake on DATA in
the "noerror" part of the transfer routine, because again the
device does not send it explicitly.