Import IEEE-488 code from Thomas Winkler (t.winkler@tirol.com) I did some reworking to integraate with IEC routines.
IEC/IEEE both have a struct ProtocolFunctions that has pointers to
their routines. cbm_init() runs both iec_init() and ieee_init() and returns
whichever one claims the bus.

Also add support for announcing the device is in IEEE mode in response to
the XUM1541_INIT command.

Auto-detect bus type when we receive the XUM1541_INIT msg. That way
the user has time to turn on either the IEC or IEEE drive after plugging
the xum1541 device into USB. We then auto-detect and stick to that
determination of the active bus until power off.