Remove libcbmcopy and libd64copy from opencbm.so
libcbmcopy and libd64copy were put into opencbm.so so 3rd party tools
can access them. AFAIK, no tool ever used this possibility.
As the shared lib becomes very huge because of this, remove them and
compile them as part of cbmcopy and d64copy, respectively, as it is also
done on Windows.