Fixes for OpenCBM libusb link-order problem
The problem is quite generic and it is pure luck that is hasn't occoured long time ago.

The problem is (as always) related to binutils-ld.
Obejct files that refer to library based functions should always be
placed before the actual library on the commandline. This was not the
case with OpenCBM. The binutils-ld linker have always been quite picky
about placement of libraries - but now that ubuntu have gone from
binutils 2.21.x to 2.22.x it has become more picky than normal.

This patch fixes the above mentioned link-order problem for both xum1541 *and* xu1541 plugins.

Reports on the cbm-hacker list indicates that xu1541 has the same problem