Linux: Report success in wait_for_free_bus()
After resetting the IEC bus, the driver waits for the device(s) to
release the IEC bus. In case of an error, the error is logged. However,
in case of success, there was no log until now. This way, if someone
posted a log, it was unclear if the wait ended successfully, or if
someone forgot to add all relevant lines.

This patch adds a positive log line, making it easier to reveal the
outcome of the operation.