Add -c latin to linuxdoc/sgmltools.
Patch from Frédéric Brière from 2007-12-02, adjusted to current build
system:

Here's a patch that adds the -c <charset> option to all sgml2* commands
in docs/, so that non-ASCII characters are not (or rather, less)
mangled. It's far from perfect (the lack of UTF-8 support sucks), but
it's the best that can be done.