Patch from Frédéric Brière from 2007-12-02:
Somewhere between 4.8 and 4.11, makeinfo stopped putting leading spaces
in front of "\input texinfo", thus breaking the regex in
docs/LINUX/Makefile.

I took this as an opportunity to be a little bit bolder, and actually
replaced the perl invocation with sed instead, as this was the only part
that required perl to be installed. Feel free to disagree, in which case
you can simply add a "\s*" to the perl regex and forget this patch.