updating build dependencies for debian packaging
Adding missing packages that are needed as build dependency when running pbuilder to do dpkg-buildpackage in a clean environment
Maybe I have overdone it a bit with libusb packages. But at the moment it builds fine in pbuilder and on my PPA on launchpad.net.
I will refine this list later.