Remove jitter in iec_srq_write() that caused errors on new compilers
The previous version of this code had a conditional branch on the high bit
of the data byte being set. This determined which lines we set/cleared.

On newer compilers like gcc 8.2.0 and 9.1.0, some optimization levels
lifted one of the conditions out of the loop, resulting in an extra 4 cycles
of jitter, based on the data bit. This was enough to cause errors since
the SRQ clock coming into the drive needs to be more stable so data would
be lost.

Instead, make this loop body constant-time by removing branches.
At 500 Kbit/sec, each loop iteration should take 2 us or 32 clocks per
bit at 16 MHz. The new code is faster, taking only 13 clocks total for
the loop body. This means we have 19 clocks to use for delays.