xum1541_close(): Ignore if libusb_release_interface() fails
xum1541_close() is also called from xum1541_device_path() which
never claims the interface, so its call to xum1541_close() always
results in an error.

xum1541_init() tries to use the variable 'interface_claimed' to
keep track if the interface was claimed or not, but calls
xum1541_close() on error, which unconditionally calls
libusb_release_interface() anyway.

As this happens on close and there's nothing we could do anyway
it's save to just ignore the error.