xum_1541_close(): Ignore errors on deconfigure
libusb1 on windows returns a LIBUSB_ERROR_INVALID_PARAM error
when calling libusb_set_configuration() with a configuration
value of -1 to deconfigure the device, even though the docs
for libusb_set_configuration() mention this as a valid value:

"A configuration value of -1 will put the device in
unconfigured state. The USB specifications state that a
configuration value of 0 does this, however buggy devices
exist which actually have a configuration 0."

(A configuration value of 0 doesn't work on windows, neither).

Deconfiguring the device is necessary on linux, to leave the
device in a known state after running a command.

As deconfiguration seems to always fail on windows and even
on linux there is nothing we can really do when it fails,
just ignore a possible error in order not to confuse users.