xum1541_init(): explicitly set interface to reset data toggles
Different approach to 3b4c843 to solve the problem that in some
configurations eg "cbmctrl status" only works once after the
device was plugged in.

The reason for the failure is that the data toggles used for the
bulk transfers get out of sync, because while the xum1541 firmware
resets them every time a "set configuration" control message is
sent, that apparently isn't always true for the host.

De-configuring the device in xum1541_close() (the method used by
3b4c843) also worked, but turned out to have various negative side
effects:

- Trying to de-configuring the device with a configuration value
  of -1 - as suggested by the libusb docs - results in a
  LIBUSB_ERROR_INVALID_PARAM error (which can be ignored).
- On MacOS libusb version 1.0.23 has an assert() check in
  libusb_set_configuration() that triggers on a configuration value
  of -1. This check has since been removed, but as of this writing
  1.0.23 is still the latest released version.
- De-configuring the device seems to have negative effects on other
  applications that use the device in parallel. A WebUSB sample
  using chromium 83 wasn't able to access the xum1541 after parallel
  use of a cbmctrl version that de-configures the device in
  xum1541_close().

This patch is purely based on the evidence that with the (redundant)
call to set the interface "cbmctrl status" works on all host
configurations tested so far and the hope that it has less negative
side effects.