2 pass for 1571 in order to start at page boundary
As the 1571 does not have enough memory, I tried to put both buffers
result and result2 into the memory, not starting at a page boundary. But
this changes the timing.

Because of this, changed the implementation so that the 1571 does it in
two passes: In the first pass, result is built; in the second pass,
result2 is built.