Fix for the lockup problem of PROMICRO with SRQ
With SRQ nibbling, the PROMICRO (non-7406 variant) locks up completely.
Afterwards, it will not work at all! One has to power cycle the floppy
as well as the PRO MICRO to make it work again.

This is fixed with this patch.

Note, however, that SRQ nibbling is broken, as it is with the
PROMICRO_7406 variant. This remains to be done.