xum1541: Include details of compilation environment in firmware
Adds git revision, avr-gcc version and avr-libc version used.

New USB control messages allow to query these values. The xum1541
plugin checks in xum1541_init() if these messages are supported by
the firmware and if so prints the reported values to the debug log.

This should in the long run help to diagnose problems reported by
users.

If the source isn't in a local git repository or the git command
is not found, the git revision is reported as "unknown".