Set LUFA's NO_LIMITED_CONTROLLER_CONNECT build flag
Even though it no longer keeps the device from reconnecting after a
USB suspend message, the connect / disconnect event fallback still
makes the firmware enter the disconnected state (STATUS_INIT).

This turns on the LED, which isn't really useful, as it's not
relevant for the user.

So set the NO_LIMITED_CONTROLLER_CONNECT build flag to disable the
event fallback.