Remove the XUM1541_NO_DEVICE status flag
It's not used anywhere in the opencbm host tools plus the comments in
the code are misleading: "device_running" has nothing to do with IEC
devices, but indicates if the xum1541 device is connected to USB (it's
set in the "configuration changed" event handler in main.c).

So the flag doesn't indicate what it claims it does, it's always set
because the XUM1541_INIT command wouldn't have been received if USB
wasn't connected and it isn't used, so get rid of it altogether.