d64copy writes errormap with --bam--* options
The options --bam-only (-b) or --bam-save (-B) make d64copy write out an
error map, even if no errors occurred.

The analysis showed the following reason: The error_map[] is initialized
with all "0". Whenever a block is read, the corresponding block is
written, and the error_map[] entry gets the result of the operation. If
it is successfull, it will be 1.

After everything is read, d64copy checks if an error occurred; for this,
it checks all entrys in error map if they are 1. Unfortunately, for all
the blocks that were skipped, they were 0, so the error information is
erroneously written to disk.

The fix is the ignore error codes of 0 when checking if an error
occurred.

Note that for the 1581 and CMD drives, both the error codes "0" and "1"
mean success. For older drives like the 154x and 157x, the error code 0
is not even defined. Thus, adding "0" is success does not do any harm
on these older drives.

Note, however, bug #116 which adds to this discussion. After thinking
about it, I am not sure if "fixing" this issue is a good idea in the
first place.

This same problem also occurs with imgcopy and d82copy.
(And this also prooves that it is a nightmare to support three tools
that are so similar)

This fixes #115. Thanks to github user "movvny" for reporting this
issue.