xum1541cfg: Compile for C99 standard
xum1541cfg uses DFU-Programmer in a rather old version (0.5.4). It is
not built-compatible with C23 anymore, because it defines true and
false.

This results in build failures on gcc15 and newer. This patch fixes it
by forcing the C standard to c99 for xum1541cfg (only). Other than #140,
this is not enforced on the whole code base of OpenCBM.

This should fix #139 and be less invasive than #140.