Do not use usleep() anymore
usleep() has been deprecated a long time ago by POSIX.1-2001. It should
be replaced with nanosleep().

Newer gcc does not include it anymore by default. This patch uses
nanosleep() instead. Additionally, it changes the way the functions
arch_sleep() and arch_usleep() are implemented on Windows and Linux.
Additionally, there are new functions arch_sleep_s(), arch_sleep_ms()
and arch_sleep_us() to sleep for seconds, milliseconds or microseconds,
resp.

Note also that the resolution of sleep on Windows is limited to
multiples of milliseconds to begin with. Only timeBeginTime() refines
the resolution to 1 ms for OpenCBM, but this is only done with the
XA1541 plugin. Thus, all timeouts on Windows must be handled with care!